chore: Update pyproject.toml to disable package mode
The pyproject.toml file was updated to set the package-mode flag to false. This change was made to disable the package mode for the project.

chore: Update .vscode/settings.json to change BASE_URL

The .vscode/settings.json file was updated to change the BASE_URL value. The previous value was commented out and a new value was added. This change was made to update the base URL for the project.

chore: Update complaint_untreated.py, complaint_treat.py, suggestion_untreated.py, suggestion_treat.py, and sidebar.py to comment out UI notifications

Several files were updated to comment out UI notifications. The lines that contain the "ui.notify" function were commented out. This change was made to temporarily disable the UI notifications.

chore: Update sidebar.py to add "导出数据" item

The sidebar.py file was updated to add a new item called "导出数据". This change was made to add a new menu item for exporting data.

chore: Update pim.py to comment out get_excel import and download_excel function

The pim.py file was updated to comment out the import statement for the get_excel function and the download_excel function. This change was made to temporarily disable the functionality related to exporting Excel files.